fix: resolve pkg-file-count and UI coverage mergeability failures
