+++ docs/internal/processes/api-relevant-files.md
@@ -16,9 +16,9 @@ Use this map when changing the public REST contract.
 - Orchestrator-aware factory events live in `api/components/schemas/events/payloads/JavaScriptPhaseChangeEventPayload.yaml` and `ArtifactCreatedEventPayload.yaml`; keep Petri marking progress on `WORK_STATE_CHANGE`, JavaScript workflow progress on `JAVASCRIPT_PHASE_CHANGE`, and replay orchestrator projections through `pkg/factory/projections/world_state.go` plus `pkg/factory/projections/world_view_runtime.go`.
 - Session lifecycle factory events live in `api/components/schemas/events/payloads/SessionStartedEventPayload.yaml`, `SessionResultUpdatedEventPayload.yaml`, `SessionCompletedEventPayload.yaml`, `OrchestratorPhaseChangedEventPayload.yaml`, `OrchestratorCheckpointWrittenEventPayload.yaml`, `DispatchQueuedEventPayload.yaml`, `DispatchInterruptedEventPayload.yaml`, and `DispatchReconciledEventPayload.yaml`; keep shared identity and ordering on `FactoryEventContext` (`sessionId`, `sessionSequence`, `orchestratorKind`, `orchestratorDialect`, `phaseId`, `phaseName`, `checkpointId`, `dispatchId`), use `FactoryEventSessionResultStatus` for event payloads (distinct from durable read-model `FactorySessionResultStatus`), and prove contract parity through `pkg/api/contracttests/` plus `ui/src/api/events/types.test.ts`.
 - Session bracket runtime emission and replay live in `pkg/factory/events/event_history_session_lifecycle.go`, `pkg/factory/runtime/factory.go`, and `pkg/factory/projections/world_state_session_lifecycle.go`; prove successful and failed-with-partial reconstructability through `pkg/factory/projections/projectiontests/session_lifecycle_event_replay_test.go` and `pkg/factory/events/event_history_session_lifecycle_test.go`.
-- Orchestrator phase/checkpoint runtime emission and replay live in `pkg/factory/events/event_history_orchestrator_progress.go` and `pkg/factory/projections/world_state_orchestrator_progress.go`; prove phase history, latest checkpoint ref, and checkpoint warnings through `pkg/factory/projections/projectiontests/orchestrator_event_replay_test.go` and `pkg/factory/events/event_history_orchestrator_progress_test.go`.
-- Dispatch queue/interruption/reconciliation and artifact-created runtime emission and replay live in `pkg/factory/events/event_history_dispatch_lifecycle.go` and `pkg/factory/projections/world_state_dispatch_lifecycle.go`; prove JavaScript dispatch lifecycle plus Petri dispatch-request compatibility through `pkg/factory/projections/projectiontests/dispatch_lifecycle_event_replay_test.go` and `pkg/factory/events/event_history_dispatch_lifecycle_test.go`.
-- Event stream reconnect lives in `api/components/parameters/AfterEventId.yaml`, `api/components/parameters/AfterSequence.yaml`, `pkg/factory/events/event_reconnect.go`, and `pkg/api/handlers_events.go`; prove cursor filtering and reconnect replay reconstruction through `pkg/factory/events/event_reconnect_test.go`, `pkg/factory/projections/projectiontests/session_reconnect_replay_test.go`, and `pkg/api/servertests/server_dashboard_events_test.go`.
+- Orchestrator phase/checkpoint runtime emission lives in `pkg/factory/events/event_history_orchestrator_progress.go`; replay projection lives in `pkg/factory/projections/world_state_session_lifecycle.go` alongside session bracket reducers; prove phase history, latest checkpoint ref, and checkpoint warnings through `pkg/factory/projections/projectiontests/orchestrator_event_replay_test.go` and `pkg/factory/events/event_history_orchestrator_progress_test.go`.
+- Dispatch queue/interruption/reconciliation, artifact-created runtime emission, and stream reconnect replay live in `pkg/factory/events/event_history_dispatch_lifecycle.go` (`BuildReconnectReplay`); dispatch lifecycle replay projection lives in `pkg/factory/projections/world_state_dispatch.go`; prove JavaScript dispatch lifecycle plus Petri dispatch-request compatibility through `pkg/factory/projections/projectiontests/dispatch_lifecycle_event_replay_test.go` and `pkg/factory/events/event_history_dispatch_lifecycle_test.go`.
+- Event stream reconnect query params live in `api/components/parameters/AfterEventId.yaml`, `api/components/parameters/AfterSequence.yaml`, and `pkg/api/handlers_events.go`; prove cursor filtering and reconnect replay reconstruction through `pkg/factory/events/event_reconnect_test.go`, `pkg/factory/projections/projectiontests/session_reconnect_replay_test.go`, and `pkg/api/servertests/server_dashboard_events_test.go`.
 - Cross-surface lifecycle exposure lives in `ui/src/api/events/api.ts` (reconnect cursors), `ui/src/features/dashboard/hooks/event-stream/useFactoryEventStream.ts`, `ui/src/features/timeline/state/timeline/replaySessionLifecycle.ts`, `ui/src/features/timeline/state/timeline/replayOrchestratorProgress.ts`, `ui/src/features/timeline/state/timeline/replayDispatchLifecycle.ts`, `ui/src/features/dashboard/components/dashboard-session-lifecycle-banner.tsx`, and `pkg/cli/session/show.go`; document API/CLI/dashboard/MCP semantics in `docs/reference/sessions.md`.
 - Orchestrator-aware public surfaces: extend `ui/src/api/factory-sessions/api.ts` with `getFactorySession` / result reads, render dashboard runtime detail through `ui/src/features/factory-session-detail/`, and expose the `factory-session` bento widget from `ui/src/features/bento/hooks/dashboardLayoutSchema.ts`.
 - Shared dispatch and artifact runtime projections live in `api/components/schemas/data-models/FactoryDispatch*.yaml` and `api/components/schemas/data-models/FactoryArtifact.yaml`; map Petri dispatches from `pkg/factorysessions/dispatch_projection.go` via engine snapshots, JavaScript dispatches/artifacts from `pkg/interfaces/factory_dispatch_projection.go`, and expose both through `FactorySessionRuntime.dispatches` / `artifacts` in `pkg/factorysessions/projection.go`.
